Updated file copying to use jinja2.
File copying now uses template render to actual get the contents of the
file to output. Since there is no actual templating being done (yet)
this means it's just a fancy way of copying files for now.

Also, I had to split up the files that are input to this new "copying"
process into those that are text-based (so files that would potentially
be templates in the future) and those that are not (e.g. images). Had to
do this because it's not possible to use template render on binary
images to get their output. So files in the static folder just use the
normal copy into the output folder.