Recursive copying of the static folder.
Previously copying of the static folder would only copy the top level of
static/ to the output folder. This isn't ideal when we want to copy
nested sub directories from static/ to output/. In fact, adding a folder
to the static/ folder broke the previous copying routine because it
assumed everything in the folder was a file.

I did try to write my own recursive copying function but abandoned that
as it turns out there is a built method for doing so. May need to
revisit a custom recursive copier in the future though, for when the
templates and content folder have nested sub directories.