Added data injection from json.
The json data file is indexed by file name and relates to an array of
objects that are 'template variable':'value'. Now it's possible for the
same template variable to have different values for different files.

To make this work, the dictionary that gets passed to the template
render method is a concatenation of the original 'static' content
dictionary (static in so much as the injected content is the same for
each page) and the new dynamic values (per page).