Rewrote existing scaffss builder to use config in json.
The previous method for building a file has now been translated into
using the config as defined by the scaffss.json file. This json gets
deserialised into an object which contains a list of pages, each page
has it's file location (so what was previsouly referred to a template),
the additional files that get injected into the page and the literals
that get inject into the page.