Tidying up.
Removing the dump.txt file that I was using as a holding file,
effectively for notes for the scaffss.json file I was building.

Also, removing the old copy_folder example we started with at the
beginning of the project. It's not needed anymore as scaffss uses a more
sophisticated method for generating the output and adding unnecessary
bloat to the code. Some removing it now will make it easier to tidy up
the actual scaffss code in the future.