Updated the scaffss script to take the scaffss location as a cmd line arg.
This way the location of the script to process isn't baked into the
scaffss script anymore. So it's much more generic.

I've also updated the scaffss script so it finds the directory of the
scaffss file and uses this as a root directory for all file locations
referenced in the scaffss file. This means all the file locations in the
scaffss file can now be relative to the file itself, meaning lot of the
repeated path info that was present in the scaffss before has now been
removed.